move AssistantUrlDomainAnalysisResults.jsx to the AssistantCheckResults folder
